move DUCC from pages/profile to components
